return of the warnings

+++ arm9/lib/src/GX_bgcnt.c
@@ -3,7 +3,7 @@
 #include "GX.h"
 
 ARM_FUNC void *G2_GetBG0ScrPtr(){
-    u32 temp = (((reg_G2_BG0CNT & 0x1F00) >> 0x8) << 0xB);
+    u32 temp = (u32)(((reg_G2_BG0CNT & 0x1F00) >> 0x8) << 0xB);
     return (void *)(0x6000000 + (((reg_GX_DISPCNT & 0x38000000) >> 0x1B) << 0x10) + temp);
 }
 
@@ -12,7 +12,7 @@ ARM_FUNC void *G2S_GetBG0ScrPtr(){
 }
 
 ARM_FUNC void *G2_GetBG1ScrPtr(){
-    u32 temp = (((reg_G2_BG1CNT & 0x1F00) >> 0x8) << 0xB);
+    u32 temp = (u32)(((reg_G2_BG1CNT & 0x1F00) >> 0x8) << 0xB);
     return (void *)(0x6000000 + (((reg_GX_DISPCNT & 0x38000000) >> 0x1B) << 0x10) + temp);
 }
 
@@ -119,7 +119,7 @@ ARM_FUNC void *G2S_GetBG3ScrPtr(){
 }
 
 ARM_FUNC void *G2_GetBG0CharPtr(){
-    u32 temp = (((reg_G2_BG0CNT & 0x3C) >> 0x2) << 0xE);
+    u32 temp = (u32)(((reg_G2_BG0CNT & 0x3C) >> 0x2) << 0xE);
     return (void *)(0x6000000 + (((reg_GX_DISPCNT & 0x7000000) >> 0x18) << 0x10) + temp);
 }
 
@@ -128,7 +128,7 @@ ARM_FUNC void *G2S_GetBG0CharPtr(){
 }
 
 ARM_FUNC void *G2_GetBG1CharPtr(){
-    u32 temp = (((reg_G2_BG1CNT & 0x3C) >> 0x2) << 0xE);
+    u32 temp = (u32)(((reg_G2_BG1CNT & 0x3C) >> 0x2) << 0xE);
     return (void *)(0x6000000 + (((reg_GX_DISPCNT & 0x7000000) >> 0x18) << 0x10) + temp);
 }
 
@@ -137,7 +137,7 @@ ARM_FUNC void *G2S_GetBG1CharPtr(){
 }
 
 ARM_FUNC void *G2_GetBG2CharPtr(){
-    s32 temp1 = (reg_GX_DISPCNT & 0x7);
+    s32 temp1 = (s32)(reg_GX_DISPCNT & 0x7);
     u32 temp = reg_G2_BG2CNT;
     if (temp1 < 5 || !(temp & 0x80))
     {
@@ -152,7 +152,7 @@ ARM_FUNC void *G2_GetBG2CharPtr(){
 }
 
 ARM_FUNC void *G2S_GetBG2CharPtr(){
-    s32 temp1 = (reg_GXS_DB_DISPCNT & 0x7);
+    s32 temp1 = (s32)(reg_GXS_DB_DISPCNT & 0x7);
     u32 temp = reg_G2S_DB_BG2CNT;
     if (temp1 < 5 || !(temp & 0x80))
     {
@@ -166,7 +166,7 @@ ARM_FUNC void *G2S_GetBG2CharPtr(){
 }
 
 ARM_FUNC void *G2_GetBG3CharPtr(){
-    s32 temp1 = (reg_GX_DISPCNT & 0x7);
+    s32 temp1 = (s32)(reg_GX_DISPCNT & 0x7);
     u32 temp = reg_G2_BG3CNT;
     if (temp1 < 3 || (temp1 < 6 && !(temp & 0x80)))
     {
@@ -181,7 +181,7 @@ ARM_FUNC void *G2_GetBG3CharPtr(){
 }
 
 ARM_FUNC void *G2S_GetBG3CharPtr(){
-    s32 temp1 = (reg_GXS_DB_DISPCNT & 0x7);
+    s32 temp1 = (s32)(reg_GXS_DB_DISPCNT & 0x7);
     u32 temp = reg_G2S_DB_BG3CNT;
     if (temp1 < 3 || (temp1 < 6 && !(temp & 0x80)))
     {
